2293: - resolving merge conficts ...
